Move domain parameter to user command
A domain can now be specified for the user command, in favour of
specifying it in each user subcommand (edit, add, etc.). Further,
a default domain can be specified in the .env file using
DOMAIN_DEFAULT. The domain parameter for the user command (-d|--domain)
however has precedence over the .env configuration.