Use DatabaseConnection everywhere in vmail-lib
While the type DatabaseConnection is hiding the actual database backend
for access from outside the library, internal files still used an
explicit reference to the default mysql backend, instead of using the
public type DatabaseConnection.

This commit extracts the connect function into a separate module which
also defines the used backend. This allows to switch to another
database backend much easier, since only Cargo.toml (selecting the
diesel features) and one line in src/database.rs have to be changed.
The behaviour of vmail-lib is unchanged.