Support adding alias forwarding to external domain
It is great to check that the alias destination is a valid address
on the server, this actually prevents some mistakes.
But in some cases, you actually want to forward to an external domain,
for example user.name@mydomain.org => user.name@gmail.com
The destination validity check currently prevents this.

This patch allows explicitly bypassing the destination validity check
to allow adding alias forwarding to external domain using the --forward
argument.